requirements.txt
anthropic>=0.40.0
pydantic>=2.6.0
openpyxl>=3.1.2
rich>=13.7.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
streamlit>=1.33.0
httpx>=0.27.0
python-multipart>=0.0.9
